Refactor: Move NestedDoStatement to cx2x.xcodeml.helper
